Removed storage display and fixed inline expr bug
The display method has been removed from the variable storage and also the runner.
While testing, I found an issue with inline expressions, one of today's revisions was the cause. I fixed it. It had to do with display a variable that had a numeric value of 0, it was showing "null" instead.